So I looked at the code last time, and it just looked to squished together and messy and I had no idea where to go with any of it.
But, I did look at a video and they implemented a Position class. So I implemented a Position class to hold all the values for the snake. I think this way, the code will be much more clearer to read AND understand. I'm feeling pretty confident with this new change.